[FEAT] 비동기 처리 순서 수정
