+++ src/main/java/com/ll/playon/domain/chat/controller/ChatController.java
@@ -30,7 +30,6 @@ public class ChatController {
         // TODO: 배포 시 주석 해제, @RequestHeader 삭제
 //        Member actor = this.userContext.getActualActor();
         Member actor = this.userContext.findById(userId);
-        System.out.println("ACTOR 정보 : " + actor.getId());
 
         return RsData.success(HttpStatus.OK, this.chatService.enterPartyRoom(actor, partyId));
     }
