+++ src/main/java/com/ll/playon/domain/member/service/MemberService.java
@@ -4,14 +4,10 @@ import com.ll.playon.domain.game.game.dto.GameListResponse;
 import com.ll.playon.domain.game.game.entity.SteamGame;
 import com.ll.playon.domain.game.game.entity.SteamGenre;
 import com.ll.playon.domain.game.game.repository.GameRepository;
+import com.ll.playon.domain.game.game.repository.WeeklyGameRepository;
 import com.ll.playon.domain.game.game.service.GameService;
 import com.ll.playon.domain.image.type.ImageType;
-import com.ll.playon.domain.member.dto.GetMembersResponse;
-import com.ll.playon.domain.member.dto.MemberDetailDto;
-import com.ll.playon.domain.member.dto.MemberProfileResponse;
-import com.ll.playon.domain.member.dto.PresignedUrlResponse;
-import com.ll.playon.domain.member.dto.ProfileMemberDetailDto;
-import com.ll.playon.domain.member.dto.PutMemberDetailDto;
+import com.ll.playon.domain.member.dto.*;
 import com.ll.playon.domain.member.entity.Member;
 import com.ll.playon.domain.member.entity.MemberSteamData;
 import com.ll.playon.domain.member.entity.enums.Role;
@@ -27,17 +23,19 @@ import com.ll.playon.global.aws.s3.S3Service;
 import com.ll.playon.global.exceptions.ErrorCode;
 import com.ll.playon.global.security.UserContext;
 import com.ll.playon.global.steamAPI.SteamAPI;
-import java.time.LocalDateTime;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.UUID;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.ObjectUtils;
+
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
@@ -55,6 +53,7 @@ public class MemberService {
     private final S3Service s3Service;
     private final WeeklyGameRepository weeklyGameRepository;
     private final SteamAsyncService steamAsyncService;
+    private final ApplicationEventPublisher publisher;
 
     public Optional<Member> findById(Long id) {
         return memberRepository.findById(id);
@@ -152,7 +151,7 @@ public class MemberService {
                 .build();
         memberRepository.saveAndFlush(newMember);
 
-        steamAsyncService.getUserGamesAndCheckGenres(newMember);
+        publisher.publishEvent(new SignupEvent(newMember));
 
         return newMember;
     }
@@ -190,27 +189,7 @@ public class MemberService {
                 .orElseThrow(ErrorCode.AUTHORIZATION_FAILED::throwServiceException);
     }
 
-    public void getUserGamesAndCheckGenres(Member actor) {
-        Member member = memberRepository.findById(actor.getId())
-                .orElseThrow(ErrorCode.MEMBER_NOT_FOUND::throwServiceException);
-
-        int before = member.getGames().size();
-
-        List<Long> userGames = steamAPI.getUserGames(member.getSteamId());
-        if (!userGames.isEmpty()) {
-            SteamGenre preferredGenre = steamAPI.getPreferredGenre(userGames);
-            memberRepository.save(member.toBuilder().preferredGenre(preferredGenre.getName()).build());
-            saveUserGameList(userGames, member);
-        }
-
-        int after = userGames.size();
-
-        // 스팀 게임 소유 칭호
-        titleEvaluator.gameCountCheck(ConditionType.STEAM_GAME_COUNT, member, after - before);
-    }
-
     public PresignedUrlResponse modifyMember(PutMemberDetailDto req, Member actor) {
-    public void modifyMember(MemberDetailDto req, Member actor) {
         Member member = memberRepository.findById(actor.getId())
                 .orElseThrow(ErrorCode.AUTHORIZATION_FAILED::throwServiceException);
