+++ src/main/java/com/ll/playon/domain/member/service/SignupEvent.java
@@ -1,0 +1,7 @@
+package com.ll.playon.domain.member.service;
+
+import com.ll.playon.domain.member.entity.Member;
+
+public record SignupEvent(
+        Member member
+) {}
